pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ssp-toolkit"
version = "3.0.1"
description = "Scripts for managing the creation of an SSP"
readme = "README.md"
requires-python = ">=3.14"
license = "GPL-3.0 AND CC0-1.0"
keywords = ["ssp", "security", "compliance", "automation"]
authors = [
    {name = "Tom Camp", email = "tom.camp@civicactions.com"},
    {name = "Fen Lebalme", email = "fen@civicactions.com"},
    {name = "Alanna Blinn", email = "alanna.blinn@civicactions.com"},
]
dependencies = [
    "blinker>=1.9.0",
    "click>=8.3.1",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "markdown>=3.10",
    "md-toc>=9.0.0",
    "pydantic>=2.12.5",
    "pypandoc>=1.16.2",
    "python-dotenv>=1.2.1",
    "ruamel-yaml>=0.19.1",
    "python-slugify>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.18.2",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]

[project.urls]
Source = "https://github.com/CivicActions/ssp-toolkit"

[project.scripts]
cli = "tools.cli:cli"

[[tool.mypy.overrides]]
module = ["yaml"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tools.export_to"]
disable_error_code = ["import-untyped"]

[tool.isort]
profile = "black"
line_length = 109

[tool.black]
line-length = 109

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "types-PyYAML",
]
